requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0

# ORM
SQLAlchemy==2.0.23

# PostgreSQL driver
psycopg[binary]==3.2.1
psycopg2-binary==2.9.9

# Env loader
python-dotenv==1.0.0

# JWT
python-jose==3.3.0

# Password hashing
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# Pydantic (compatible with FastAPI 0.100 and Windows, no Rust)
pydantic==2.7.4
pydantic-core==2.18.4
pydantic-settings==2.2.1

# Multipart (file uploads)
python-multipart==0.0.20

# Migrations
alembic==1.12.0


#pagination
fastapi-pagination